# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Web-Shiori API document"
  version: "1.0.0"
  title: "Web-Shiori API"
host: "web-shiori-api-document.herokuapp.com"
...
schemes:
- "https"
paths:
...
  /v1/folder:
    get:
      tags:
      - "folder"
      summary: "フォルダ一覧取得API"
      description: ""
      operationId: "v1FolderGET"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "access-token"
        in: "header"
        description: "アクセストークン"
        required: true
        type: "string"
      - name: "client"
        in: "header"
        description: "client"
        required: true
        type: "string"
      - name: "uid"
        in: "header"
        description: "uid"
        required: true
        type: "string"
      responses:
        "200":
          description: "A successful response."
          schema:
            $ref: "#/definitions/FolderListResponse"
      x-swagger-router-controller: "Folder"
    post:
      tags:
      - "folder"
      summary: "フォルダ作成API"
      description: ""
      operationId: "v1FolderPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "access-token"
        in: "header"
        description: "アクセストークン"
        required: true
        type: "string"
      - name: "client"
        in: "header"
        description: "client"
        required: true
        type: "string"
      - name: "uid"
        in: "header"
        description: "uid"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Request Body"
        required: true
        schema:
          $ref: "#/definitions/FolderCreateRequest"
      responses:
        "200":
          description: "A successful response."
          schema:
            $ref: "#/definitions/FolderCreateResponse"
      x-swagger-router-controller: "Folder"
  /v1/folder/{folder_id}:
    put:
      tags:
      - "folder"
      summary: "フォルダ情報更新API"
      description: ""
      operationId: "v1FolderFolder_idPUT"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "folder_id"
        in: "path"
        required: true
        type: "integer"
      - name: "access-token"
        in: "header"
        description: "アクセストークン"
        required: true
        type: "string"
      - name: "client"
        in: "header"
        description: "client"
        required: true
        type: "string"
      - name: "uid"
        in: "header"
        description: "uid"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Request Body"
        required: true
        schema:
          $ref: "#/definitions/FolderUpdateRequest"
      responses:
        "200":
          description: "A successful response."
          schema:
            $ref: "#/definitions/FolderUpdateResponse"
      x-swagger-router-controller: "Folder"
    delete:
      tags:
      - "folder"
      summary: "フォルダ削除API"
      description: ""
      operationId: "v1FolderFolder_idDELETE"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "folder_id"
        in: "path"
        required: true
        type: "integer"
      - name: "access-token"
        in: "header"
        description: "アクセストークン"
        required: true
        type: "string"
      - name: "client"
        in: "header"
        description: "client"
        required: true
        type: "string"
      - name: "uid"
        in: "header"
        description: "uid"
        required: true
        type: "string"
      responses:
        "204":
          description: "A successful response."
      x-swagger-router-controller: "Folder"
  /v1/folder/{folder_id}/content:
    get:
      tags:
      - "folder"
      summary: "フォルダ内コンテンツ一覧取得API"
      description: ""
      operationId: "v1FolderFolder_idContentGET"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "folder_id"
        in: "path"
        required: true
        type: "integer"
      - name: "access-token"
        in: "header"
        description: "アクセストークン"
        required: true
        type: "string"
      - name: "client"
        in: "header"
        description: "client"
        required: true
        type: "string"
      - name: "uid"
        in: "header"
        description: "uid"
        required: true
        type: "string"
      - name: "q"
        in: "query"
        description: "検索クエリ"
        required: false
        type: "string"
      - name: "limit"
        in: "query"
        description: "取得数。デフォルトは50"
        required: false
        type: "integer"
      - name: "offset"
        in: "query"
        description: "取得位置。デフォルトは0"
        required: false
        type: "integer"
      responses:
        "200":
          description: "A successful response."
          schema:
            $ref: "#/definitions/FolderContentGetResponse"
      x-swagger-router-controller: "Folder"
  /v1/folder/{folder_id}/content/{content_id}:
    post:
      tags:
      - "folder"
      summary: "フォルダにコンテンツを追加するAPI"
      description: ""
      operationId: "v1FolderFolder_idContentContent_idPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "folder_id"
        in: "path"
        required: true
        type: "integer"
      - name: "content_id"
        in: "path"
        required: true
        type: "integer"
      - name: "access-token"
        in: "header"
        description: "アクセストークン"
        required: true
        type: "string"
      - name: "client"
        in: "header"
        description: "client"
        required: true
        type: "string"
      - name: "uid"
        in: "header"
        description: "uid"
        required: true
        type: "string"
      responses:
        "204":
          description: "A successful response."
      x-swagger-router-controller: "Folder"
    delete:
      tags:
      - "folder"
      summary: "フォルダからコンテンツを削除するAPI"
      description: ""
      operationId: "v1FolderFolder_idContentContent_idDELETE"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "folder_id"
        in: "path"
        required: true
        type: "integer"
      - name: "content_id"
        in: "path"
        required: true
        type: "integer"
      - name: "access-token"
        in: "header"
        description: "アクセストークン"
        required: true
        type: "string"
      - name: "client"
        in: "header"
        description: "client"
        required: true
        type: "string"
      - name: "uid"
        in: "header"
        description: "uid"
        required: true
        type: "string"
      responses:
        "204":
          description: "A successful response."
      x-swagger-router-controller: "Folder"
definitions:
...
  FolderListResponse:
    type: "object"
    properties:
      data:
        type: "object"
        properties:
          folder:
            type: "array"
            items:
              $ref: "#/definitions/Folder"
    example:
      data: "{}"
  FolderContentGetResponse:
    type: "object"
    properties:
      meta:
        $ref: "#/definitions/MetaData"
      data:
        type: "object"
        properties:
          folder:
            $ref: "#/definitions/Folder"
          content:
            type: "array"
            items:
              $ref: "#/definitions/Content"
    example:
      data: "{}"
      meta:
        q: ""
        offset: 0
        limit: 50
        hasNext: true
  FolderCreateRequest:
    type: "object"
    properties:
      name:
        type: "string"
        description: "フォルダ名"
    example:
      name: "name"
  FolderCreateResponse:
    type: "object"
    properties:
      data:
        $ref: "#/definitions/Folder"
    example:
      data:
        name: "動画"
        folder_id: 1
        content_count: 10
  FolderUpdateRequest:
    type: "object"
    properties:
      name:
        type: "string"
        description: "フォルダ名"
    example:
      name: "name"
  FolderUpdateResponse:
    type: "object"
    properties:
      data:
        $ref: "#/definitions/Folder"
    example:
      data:
        name: "動画"
        folder_id: 1
        content_count: 10
  MetaData:
    type: "object"
    properties:
      q:
        type: "string"
        example: ""
        description: "検索クエリ"
      limit:
        type: "integer"
        example: 50
        description: "取得数"
      offset:
        type: "integer"
        example: 0
        description: "取得位置(相対位置)"
      hasNext:
        type: "boolean"
        example: true
        description: "続きが存在するかどうか"
    example:
      q: ""
      offset: 0
      limit: 50
      hasNext: true
...
  Content:
    type: "object"
    properties:
      content_id:
        type: "integer"
        example: 126
        description: "コンテンツid"
      type:
        type: "string"
        example: "web"
        description: "type"
      title:
        type: "string"
        example: "日向坂で会いましょう"
        description: "title"
      url:
        type: "string"
        example: "https://hinatasaka.com"
        description: "url"
      sharing_url:
        type: "string"
        example: "https://universal_link.hinatasaka.com"
        description: "sharing_url"
      file_url:
        type: "string"
        example: "null"
        description: "file_url"
      scroll_position_x:
        type: "integer"
        example: 0
        description: "scroll_position_x"
      scroll_position_y:
        type: "integer"
        example: 200
        description: "scroll_position_y"
      max_scroll_position_x:
        type: "integer"
        example: 0
        description: "max_scroll_position_x"
      max_scroll_position_y:
        type: "integer"
        example: 10000
        description: "max_scroll_position_y"
      video_playback_position:
        type: "integer"
        example: 500
        description: "video_playback_position"
      specified_text:
        type: "string"
        example: "null"
        description: "specified_text"
      specified_dom_id:
        type: "string"
        example: "null"
        description: "specified_dom_id"
      specified_dom_class:
        type: "string"
        example: "null"
        description: "specified_dom_class"
      specified_dom_tag:
        type: "string"
        example: "null"
        description: "specified_dom_tag"
      liked:
        type: "boolean"
        example: true
        description: "liked"
      delete_flag:
        type: "boolean"
        example: false
        description: "削除されたかどうか。削除から30日間は復元できるようにするために保持する。"
      deleted_at:
        type: "string"
        example: "null"
        description: "deleted_at"
      created_at:
        type: "string"
        example: "2019-05-12T20:48:24.000+09:00"
        description: "created_at"
      updated_at:
        type: "string"
        example: "2019-05-12T20:48:24.000+09:00"
        description: "updated_at"
    example:
      file_url: "null"
      specified_dom_id: "null"
      content_id: 126
      max_scroll_position_y: 10000
      created_at: "2019-05-12T20:48:24.000+09:00"
      scroll_position_y: 200
      scroll_position_x: 0
      max_scroll_position_x: 0
      type: "web"
      title: "日向坂で会いましょう"
      deleted_at: "null"
      url: "https://hinatasaka.com"
      sharing_url: "https://universal_link.hinatasaka.com"
      liked: true
      video_playback_position: 500
      specified_dom_class: "null"
      updated_at: "2019-05-12T20:48:24.000+09:00"
      specified_text: "null"
      delete_flag: false
      specified_dom_tag: "null"
  Folder:
    type: "object"
    properties:
      folder_id:
        type: "integer"
        example: 1
        description: "フォルダid"
      name:
        type: "string"
        example: "動画"
        description: "フォルダ名"
      content_count:
        type: "integer"
        example: 10
        description: "フォルダに含まれているコンテンツの数"
    example:
      name: "動画"
      folder_id: 1
      content_count: 10